解决了数字溢出和类型不匹配问题

+++ Assets/scripts/UI/ShopFormulaSlot.cs
@@ -123,7 +123,7 @@ public class ShopFormulaCardSlot : MonoBehaviour
     /// <summary>
     /// 更新价格显示
     /// </summary>
-    void UpdatePriceDisplay(int price, bool sold)
+    void UpdatePriceDisplay(long price, bool sold)
     {
         if (priceText == null) return;
